+++ CHANGELOG.md
@@ -12,7 +12,8 @@
 
 ### added
 
-- linear issue view to print the issue, with --web and --app flags to open them instead, similar to gh's view commands
+- linear issue view to print the issue, with --web and --app flags to open them
+  instead, similar to gh's view commands
 
 ### changed